Module aws_sdk_servicequotas::error

Modules

access_denied_exception

See AccessDeniedException

aws_service_access_not_enabled_exception

See AwsServiceAccessNotEnabledException

dependency_access_denied_exception

See DependencyAccessDeniedException

illegal_argument_exception

See IllegalArgumentException

invalid_pagination_token_exception

See InvalidPaginationTokenException

invalid_resource_state_exception

See InvalidResourceStateException

no_available_organization_exception

See NoAvailableOrganizationException

no_such_resource_exception

See NoSuchResourceException

organization_not_in_all_features_mode_exception

See OrganizationNotInAllFeaturesModeException

quota_exceeded_exception

See QuotaExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

service_exception

See ServiceException

service_quota_template_not_in_use_exception

See ServiceQuotaTemplateNotInUseException

tag_policy_violation_exception

See TagPolicyViolationException

templates_not_available_in_region_exception

See TemplatesNotAvailableInRegionException

too_many_requests_exception

See TooManyRequestsException

too_many_tags_exception

See TooManyTagsException

Structs

AccessDeniedException

You do not have sufficient permission to perform this action.

AssociateServiceQuotaTemplateError

AwsServiceAccessNotEnabledException

The action you attempted is not allowed unless Service Access with Service Quotas is

DeleteServiceQuotaIncreaseRequestFromTemplateError

DependencyAccessDeniedException

You can't perform this action because a dependency does not have access.

DisassociateServiceQuotaTemplateError

GetAWSDefaultServiceQuotaError

GetAssociationForServiceQuotaTemplateError

GetRequestedServiceQuotaChangeError

GetServiceQuotaError

GetServiceQuotaIncreaseRequestFromTemplateError

IllegalArgumentException

Invalid input was provided.

InvalidPaginationTokenException

Invalid input was provided.

InvalidResourceStateException

The resource is in an invalid state.

ListAWSDefaultServiceQuotasError

ListRequestedServiceQuotaChangeHistoryByQuotaError

ListRequestedServiceQuotaChangeHistoryError

ListServiceQuotaIncreaseRequestsInTemplateError

ListServiceQuotasError

ListServicesError

ListTagsForResourceError

NoAvailableOrganizationException

The account making this call is not a member of an organization.

NoSuchResourceException

The specified resource does not exist.

OrganizationNotInAllFeaturesModeException

The organization that your account belongs to is not in All Features mode.

PutServiceQuotaIncreaseRequestIntoTemplateError

QuotaExceededException

You have exceeded your service quota. To perform the requested action, remove some of the

RequestServiceQuotaIncreaseError

ResourceAlreadyExistsException

The specified resource already exists.

ServiceException

Something went wrong.

ServiceQuotaTemplateNotInUseException

The quota request template is not associated with your organization.

TagPolicyViolationException

The specified tag is a reserved word and cannot be used.

TagResourceError

TemplatesNotAvailableInRegionException

The Service Quotas template is not available in this AWS Region.

TooManyRequestsException

Due to throttling, the request was denied. Slow down the rate of request calls, or request

TooManyTagsException

You've exceeded the number of tags allowed for a resource. For more information, see

UntagResourceError

Enums

AssociateServiceQuotaTemplateErrorKind

DeleteServiceQuotaIncreaseRequestFromTemplateErrorKind

DisassociateServiceQuotaTemplateErrorKind

GetAWSDefaultServiceQuotaErrorKind

GetAssociationForServiceQuotaTemplateErrorKind

GetRequestedServiceQuotaChangeErrorKind

GetServiceQuotaErrorKind

GetServiceQuotaIncreaseRequestFromTemplateErrorKind

ListAWSDefaultServiceQuotasErrorKind

ListRequestedServiceQuotaChangeHistoryByQuotaErrorKind

ListRequestedServiceQuotaChangeHistoryErrorKind

ListServiceQuotaIncreaseRequestsInTemplateErrorKind

ListServiceQuotasErrorKind

ListServicesErrorKind

ListTagsForResourceErrorKind

PutServiceQuotaIncreaseRequestIntoTemplateErrorKind

RequestServiceQuotaIncreaseErrorKind

TagResourceErrorKind

UntagResourceErrorKind