Module aws_sdk_servicequotas::model

Modules

error_reason

See ErrorReason

metric_info

See MetricInfo

quota_period

See QuotaPeriod

requested_service_quota_change

See RequestedServiceQuotaChange

service_info

See ServiceInfo

service_quota

See ServiceQuota

service_quota_increase_request_in_template

See ServiceQuotaIncreaseRequestInTemplate

tag

See Tag

Structs

ErrorReason

An error that explains why an action did not succeed.

MetricInfo

Information about the CloudWatch metric that reflects quota usage.

QuotaPeriod

Information about the quota period.

RequestedServiceQuotaChange

Information about a quota increase request.

ServiceInfo

Information about a service.

ServiceQuota

Information about a quota.

ServiceQuotaIncreaseRequestInTemplate

Information about a quota increase request.

Tag

A complex data type that contains a tag key and tag value.

Enums

ErrorCode

PeriodUnit

RequestStatus

ServiceQuotaTemplateAssociationStatus